+++ src/shared/ui/scroll-area.tsx
